author modal add

+++ app/src/main/java/cloud/ptl/boardgamecollector/db/entity/Author.java
@@ -6,7 +6,7 @@ import androidx.room.PrimaryKey;
 @Entity
 public class Author {
     @PrimaryKey(autoGenerate = true)
-    public int authorId;
+    public Long authorId;
 
     public String name;
     public String surname;
